Added validation for messages for category controller APIs

+++ spheroo-admin/src/main/java/com/qburst/spherooadmin/category/CategoryServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 /**
@@ -32,12 +33,15 @@ public class CategoryServiceImpl implements CategoryService{
 
     @Override
     public Category getCategory(Long id) {
-        return categoryRepository.getReferenceById(id);
+        if(categoryRepository.existsById(id)){
+            return categoryRepository.getReferenceById(id);
+        }
+        throw new EntityNotFoundException("No category exist with given id");
     }
 
     @Override
     @Transactional
-    public boolean updateCategoryById(Long categoryId, Category category) {
+    public String updateCategoryById(Long categoryId, Category category) {
         boolean isExist = categoryRepository.existsById(categoryId);
         if(isExist){
             category.setCategoryId(categoryId);
@@ -46,17 +50,19 @@ public class CategoryServiceImpl implements CategoryService{
             serviceChargeRepository.deleteAllById(noReferenceChargeIds);
             List<Long> noReferenceServiceIds = serviceRepository.findNullCategoryServices();
             serviceRepository.deleteAllById(noReferenceServiceIds);
-            return true;
+            return "Category Updated successfully";
         }else {
-            return false;
+            throw new EntityNotFoundException("No category exist with given id");
         }
-
     }
 
     @Override
     @Transactional
     public void deleteCategory(Long id) {
-       categoryRepository.deleteByCategoryId(id);
+        if(categoryRepository.existsById(id)){
+            categoryRepository.deleteByCategoryId(id);
+        }
+       throw new EntityNotFoundException("No category exist with given id");
     }
 
     @Override
